refactor: renamed files B2C => b2c
