+++ mpesakit/services/b2c.py
@@ -2,7 +2,7 @@
 
 from mpesakit.auth import TokenManager
 from mpesakit.http_client import HttpClient
-from mpesakit.B2C import B2C, B2CRequest, B2CResponse, B2CCommandIDType
+from mpesakit.b2c import B2C, B2CRequest, B2CResponse, B2CCommandIDType
 from mpesakit.b2c_account_top_up import (
     B2CAccountTopUp,
     B2CAccountTopUpRequest,
